requirements/base.txt
attrs==24.2.0
bcrypt==4.2.0
boto3==1.35.12
django==4.2.9
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
psycopg2-binary==2.9.9
python-decouple

